requirements.txt
django
djangorestframework
opencv-python
Pillow
tensorflow
matplotlib
keras
